requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
apache-airflow==2.6.3
pytest
pytest-mock
pyarrow
mypy
